// src/app/shared/model/userorder-detail.model.ts
export class UserOrderDetail
{
    userOrderId: string='';
    userProductId: string='';
    title: string='';
    address: string='';
    color: string='';
    size: string='';
    price: string='';
    count: string='';
    orderDate: string='';
    deliveryDate: string='';
    deliveryStatus: string='';               
}